Move exportable artifacts to a different file and no longer export runtime
