+++ src/lib/index.ts
@@ -1,2 +1,2 @@
 export * from './errors';
-export * from './runtime';
+export * from './models';
